\name{spiker-package}
\alias{spiker-package}
\alias{spiker}
\docType{package}
\title{\packageTitle{spiker}}
\description{\packageDescription{spiker}}
\section{Package Content}{\packageIndices{spiker}}
\author{\packageAuthor{spiker}}
\section{Maintainer}{\packageMaintainer{spiker}}
\keyword{package}
